memblocks_lib/pyproject.toml
[project]
name = "memblocks"
version = "0.1.0"
description = "Intelligent modular memory management system for LLMs"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "motor>=3.0",
    "qdrant-client>=1.7",
    "langchain>=0.1",
    "langchain-groq>=0.1",
    "langchain-core>=0.1",
    "httpx>=0.25",
    "requests>=2.32",
    "openinference-instrumentation-langchain>=0.1.4",
    "arize-otel>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memblocks"]
